requirements.txt
# Electrical Component Classifier - Dependencies
# Professional Flask web application for electrical component classification

# Core Framework
Flask==2.3.3
Werkzeug==2.3.7

# Machine Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1

# Data Processing
numpy==1.24.3
pandas==2.0.3

# HTTP & API
requests==2.31.0
urllib3==2.0.7

# File Handling
pathlib2==2.3.7

# Logging & Monitoring
loguru==0.7.2

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Code Quality
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Security
cryptography==41.0.4

# Optional: GPU Support (uncomment if using CUDA)
# torch==2.1.0+cu118
# torchvision==0.16.0+cu118
# --extra-index-url https://download.pytorch.org/whl/cu118

# Optional: Production Server
# gunicorn==21.2.0
# gevent==23.7.0

# Optional: Environment Management
# python-dotenv==1.0.0
